rename some class
