+++ app-center-sdk/src/main/java/com/zen/android/center/sdk/injection/component/CenterComponent.java
@@ -1,4 +1,4 @@
-package com.zen.android.center.sdk.inject.component;
+package com.zen.android.center.sdk.injection.component;
 
 import com.zen.android.center.sdk.AppCenter;
